Update W State 1 - Multi-Qubit Systems.ipynb
After the submission of this project, I explored the idea that the deterministic obtaining of a state W can be summed up as the application of a single reversible gate to a set of qubits in the ground state. Hence this mathematical appendix which presents the gate in question in its matrix form for the case of three qubits. Formulas and equations have been established and verified using the SymPy Python library.