Update CONTRIBUTORS.md
It would be a great honor to be included in this list of contributors. I support my application on the recent acceptance of the following notebook: qiskit-tutorial / appendix / more_qis / Multi-Qubit_W_States_with_Tomography.ipynb. I take this opportunity to thank the team of quiskitters for reviewing this work. In this regard, as soon as Qiskit 6.0 is online, I intend to modify the notebook to completely remove QuantumProgram. Greetings to all the team.