Add clarifications about using different gates in the terra tutorial (#347)
* Add explanation on tensor product and Qiskit basis vector ordering

This commit is the first in a series designed to provide more
explanation regarding multi-qubit gates and how they are
represented in Qiskit. This commit addresses text at the start
of the "Multi-qubit gates" section.

* Add text to the SWAP gate description

* Create a new section before Toffoli gate on "Three-qubit gates"

This commit more clearly distinguishes between two and three-qubit
gates.

* Add text to Toffoli gate section

* Add text to Fredkin gate section

Link to Wikipedia article on it, plus a short description
of its action on basis vectors.

* Fixes qubit numbering convention and adds more content on 2-qubit gates

Per review request, qubits are zero-indexed. For the discussion of
2-qubit gates, I explicitly label the qubits using Qiskit's ordering
convention, to make it easier for those using the tutorial to compute
the matrix representation of the gates.