[WIP] Add content to relaxation and decoherence tutorial (#371)
* Add more content on measurement of T2^* time

This piece still needs to be fleshed out a bit
more to explain where the oscillating term comes
from. Also, it needs to explicitly point out
that amplitude damping contributes to T2^*,
which (I think) motivates the use of the Hanh echo
to isolate dephasing-like contributions to T2^*.

* Minor gramamtical/experiment changes

* Adds a warning box to the relaxation tutorial

From discussions on the Qiskit Slack channel, I think
adding more clarity about T2 vs. T2*, and explicitly calling
out this clarity, is necessary. For this reason, I've added a
red NB ``warning" cell to call this out.

* Strike out text about Lindbladians and Markovianity

Replaces the text with a reference. (Needs to be filled in.)

* Update T2 warning box

* Add explanation of difference between T2 and T2*

* Remove Markdown hyperlinks from red div environment

The hyperlinks don't seem to be showing up on
GH's native rendering, nor NB viewer. So I'm getting
rid of them.

* Cleanup from rebase

* Do a bit more cleanup

* Remove some text formatting in T2/T2* warning box

Similar to the hyperlinks, the rendered NB doesn't
show the text formatting.