Update wigner function notebook and inline wigner code
The Wigner visualization functions are going to be removed from terra
(see Qiskit/qiskit-terra#1106 ) soon. In preparation for that this
commit migrates the code for performing the visualization into a local
python module the wigner_function.ipynb notebook can call. In addition,
this commit updates all the code in that notebook to work with a current
version of qiskit-terra. It was quite out of date and no longer
functioning when used with the latest terra version.