Clean up histogram plots in entanglement intro
When reviewing the entanglement introduction notebook there were a few
aspects of the histogram plots that were confusing when coupled with the
corresponding text explanation. The first issue was when looking at the
histograms for the single qubits of the entangled pair in isolation the
results were being displayed as 00 and 01 because there were 2 classical
bits in the circuits. This was a bit confusing in the context so to
switch it to 0 and 1 those circuits are adjusted to only have a single
classical bit so that Result.get_counts() doesn't pad the 0. The other
issue was there was one spot where 2 plot_histogram() calls were made
in a single cell which only shows the second plot. To show both plots
the cell is split into 2 so each plot is shown.