fixing binder
i think i could fix the binder issues. simply by using lowercase on the username "qiskit" and correcting the the path to (...)tutorials instead of (...)tutorial forced it to rebuild the binder image.
:)
if this alone is not enough then we may try to specify python==3.6, because one guy here (https://gitter.im/jupyterhub/binder) was able to make it work just by doing so.