Make tutorial tests run all notebooks by default and configurable
This commit updates the often forgotten about test module to actually be
runnable by default. It also makes the hard coded values for path,
kernel, and timeout configurable via env variables so people can adjust
these easily without having to hand edit code. The exlusion list option
was not added because doing that in a test module is redundant a decent
test runner will support test selection in a more robust manner and that
should be leveraged instead of exclusion list.

In addition because the notebook execution can be slow this commit also
adds a .stestr.conf file so that we can run the notebooks in parallel
(just like we do for the terra unit tests). This can be done with
'stestr run' and by default it will run as many processes as you have
cpus to run a notebook each (assuming you have stestr pip installed
along with the notebook dependencies).