Update the creating a provider notebook
The creating a provider notebook was a bit out of date. The primary
issue was that the backend configuration was missing a mandatory field,
the coupling_map. This field is added and set to None, which is the
proper value for the example here. It was also using a deprecated
function, which is also updated here.