Update parallel tools notebook for terra 0.9 (#774)
This commit fixes a few issues in the parallel when running on the
latest version of terra. The biggest issues is the
qiskit.quantum_info.synthesis.two_qubit_kak() function was renamed prior
to the 0.8 release to two_qubit_decompose() (from what I can tell
two_qubit_kak() in quantum_info synthesis was never included in a
release, although I might have missed it in my git log hunting). This
updates the call to use the proper function name. In addition because
of Qiskit/qiskit-terra#2414 all bit level access as tuples has been
deprecated and will emit a warning, to remove the warnings this changes
the access to be attributes instead.