Add transpiler logging section to using the transpiler
As part of the upcoming 0.9 terra release the logging from the
transpiler has been improved and standardized around python standard lib
logging. This tutorial adds a guide on how to leverage this in a couple
of common use cases to document how someone can configure the logging
from the transpiler.