Add missing version and license information to 2_device_noise_simulation.ipynb
Note: the whole notebook was not rerun. Thus, results from earlier
executions of this notebook that are still stored may have come from a
different version of qiskit.
I believe however that no significant changes are to be expected if the
notebook were run again with this version of qiskit.