Move aqua tutorials that depend on CPLEX to legacy_tutorials
There are 2 aqua tutorials that do not work unless CPLEX is installed.
CPLEX is too involved and difficult to install in CI. It is also
proprietary software which prohibits installing it in CI (since a CI
system can't agree to the license terms). Honestly, I'm not entirely
sure why aqua has modules in it that depend solely on having proprietary
software installed, it feels like there should be an open alternative or
if there is one already the tutorial should leverage that. To make the
docs buildable this commit moves these problematic tutorials to the
legacy directory so they're not built/run by default.