Set up CI with Azure Pipelines (#894)
* Set up CI with Azure Pipelines

This commit adds the CI configuration for running sphinx builds in CI. [skip ci]

* Add missing dep and start caching pip

* Install pandoc and graphviz

* Add cvxpy install for aer noise transformations

* Move aqua tutorials that depend on CPLEX to legacy_tutorials

There are 2 aqua tutorials that do not work unless CPLEX is installed.
CPLEX is too involved and difficult to install in CI. It is also
proprietary software which prohibits installing it in CI (since a CI
system can't agree to the license terms). Honestly, I'm not entirely
sure why aqua has modules in it that depend solely on having proprietary
software installed, it feels like there should be an open alternative or
if there is one already the tutorial should leverage that. To make the
docs buildable this commit moves these problematic tutorials to the
legacy directory so they're not built/run by default.

* Workaround terra mpl drawer bug

There is a bug in the mpl drawer in 0.19.0 (and 0.19.1) release where a
custom instruction does not draw without a label parameter set. This
will need to be fixed in terra, but in the meantime this worksaround the
bug by manually setting a label.

* Add texlive to installed packages

* Install correct texlive metapackage with qcircuit

* Correct package name again

* Add another missing binary dep

* Update qcircuit version as packaged version is too old

* Try using tlmgr instead of manually downloading qcircuit

* Try using more from tlmgr instead of apt

* Fix typo

* Use bash directly

* Adjust texlive update script usage

* Give up on tlmgr and try to manual install qcircuit again

* Fix typo

* revert aqua tutorial move and remove latex install

* Remove section of aqua notebook 2 that requires cplex

* Remove another notebook with cplex