Remove invalid stestr config
In older versions of the tutorials repo there was a unit testing harness
that would generate a test case based on executing each tutorially
successfully without any errors. When that testing harness exited stestr
could be used to parallelize the execution of the notebook tests to make
it faster. However, in recent refactors the testing harness has been
removed and instead we rely on sphinx builds to validate the notebooks
work. The .stestr.conf was not removed as part of that, so this commit
takes care of that oversight and deletes the unused test runner
configuration file.

Fixes #930