Rerun error mitigation notebook to fix failed doc builds
Since #965 merged the docs builds have been failing because the notebook
results stored in the ipynb json are invalid. This leads to an error
like:

Notebook error:
NotebookValidationError in tutorials/noise/3_measurement_error_mitigation.ipynb:
'execution_count' is a required property

Failed validating 'required' in execute_result:

On instance['cells'][66]['outputs'][0]:
{'data': {'text/html': '<h3>Version Information</h3><table><tr><th>Qiskit '
                       'Software</th><...',
          'text/plain': '<IPython.core.display.HTML object>'},
 'metadata': {},
 'output_type': 'execute_result'}

which breaks the docs build. This commit attempts to fix this by just
loading the notebook and rerunning on the cells on a fresh kernel with
the latest release of qiskit. This should hopefully address the invalid
results in the json and fix the docs builds.