Use metapackage visualization extra instead of terra's
Right now we're in a weird place because there is a terra release live
that has some breaking changes for ignis and aqua. This shouldn't be a
blocker for changes in tutorials because we should be using the
latest metapackage release but because we install visualization
dependencies using qiskit-terra[visualization] we're pulling in the
latest terra release which hasn't been included in a metapackage yet.
This causes the aqua tutorials to fail because they require the latest
release. This commit fixes this by using qiskit[visualization] which
won't pull in a terra version that isn't part of known working
metapackage.