Fix statevector snapshot access in MPS tutorial
The statevector snapshot tutorial access done in the MPS tutorials was
relying on internal implementation details (inherited from marshmallow)
for how it was accessing the statevector for printing. This causes an
issue when qiskit 0.20.0 is used because there is no marshmallow anymore.
Instead of relying on the nested internal object attr access that was
used before this uses the supported data() method to return a dict of
snapshots and pull the statevector from there. This should be support
for all versions of qiskit-terra, not just <0.15.0 and is the documented
method for retrieving snapshots.

This is the last thing needed to unblock Qiskit/qiskit#998 and finally
relaese qiskit 0.20.0.