Add Code of Conduct (#1048)
The tutorials repo was previously missing a code of conduct file for the
repository. Since IBM Quantum has rewritten the code of conduct to put in
place a stronger enforcement process and tighten up the language and the
other qiskit repos were going through and updating the contents of their
CODE_OF_CONDUCT.md. This commit takes the opportunity to add the missing
CODE_OF_CONDUCT.md with this new version.

Related to Qiskit/qiskit#1057