Pin docplex version in constraints.txt for CI job (#1082)
The recent new release of docplex has broken the released version of
aqua. There is a fix in aqua Qiskit/qiskit-aqua#1420 which was released
in 0.8.1, but this hasn't been included in a metapackage release yet as
we're waiting for bugfix releases from all the other elements in the
next couple of days. Until that is included in a metapackage release we
won't be able to use the latest version of docplex in the CI job. This
commit adds a pin on the version of docplex used in a constraints file
to always use a known working version to unblock CI.