Bump cell timeout in sphinx config
The 07_asian_barrier_spread_pricing tutorial is quite slow and often
times out in CI because running the simulation takes very close to 3
minutes in the fastest case. But, the performance of the CI nodes is
quite variable since they run on virtual machines in azure's public
cloud. So many jobs fail randomly because of timeouts on that notebook.
This commit attempts to fix this by bumping the timeout 60 seconds which
hopefully will give us enough slack time in the job to handle slow CI
nodes.