Add missing int cast for QISKIT_CELL_TIMEOUT
In #1091 we added a new env var to enable downstream CI to adjust the
per cell timeout to make things more reliable in variable performance
environments. But in that PR we neglected to cast the value of the env
var, which is always a string, to an int. This commit fixes that
oversight so the new env var is actually usable.