Pin decorator version for CI
The recent release of decorator 5.0.0 earlier today are causing CI
failures because of an internal incompatibility with the random state
decorator networkx is using for random_gnp_graph() function and these
new decorator versions. To unblock CI until either networkx or decorator
fixes the issue this commit pins the version of decorator we install via
the constraints file to the last known working version.