Revert "Update aer tutorials for 0.8 release (#1153)"
In trying to debug the reason the tutorials builds for
qiskit.org/documentation are failing, things started to fail around the
same time #1153 merged. To figure out if it was the cause this commit
temporarily reverts the changes in #1153 so we can trigger a new docs
build without it applied and see if that works.

This reverts commit 2a20bed64c69853581fce0f189ddd1d13ebba331.