Revert "Revert "Update aer tutorials for 0.8 release (#1153)" (#1158)"
This commit wasn't the source of the issue with the documentation
builds. The changes were good otherwise and we only reverted it in the
interest of testing to double check it didn't somehow break things.
Since we've confirmed it wasn't the cause this commit reverts the revert
and applies back #1153.

This reverts commit 16f83761316aa715c00e80c823f3847e74abed66.