Fix pulse simulator deprecation warnings (#1245)
The pulse simulator tutorials have been emitting deprecation warnings
about the use of the system_model positional argument on the run()
method for some time. In the just released aer 0.9.0 the deprecation was
finished and that argument was removed. This is causing ci to fail on
Qiskit/qiskit#1333 blocking the metapackage 0.30.0 release. This commit
updates the notebooks to use the current syntax for specifying the
system model which works with qiskit-aer>=0.7.0.

(cherry picked from commit 563352987b26051e809152569b25e813e9d77440)