Remove Ignis from tutorials
The qiskit-ignis project has been deprecated and supersceded by the
qiskit-experiments repository. In preparation for archiving and marking
Ignis as EoL this commit removes the tutorials explicitly showcasing
ignis or 100% dependent on ignis. Additionally, the vqe tutorial which
was leveraging ignis for measurement mitigation is refactored to use
terra's fork instead.