Update operator tutorial to avoid deprecated Pauli construction
This commit updates the operator tutorial to avoid the deprecated Pauli
constructor kwarg, label. This has been deprecated since Qiskit Terra
0.17.0 and is pending removal in Qiskit/qiskit-terra#8070. The label
kwarg isn't needed anymore as the Pauli string can just be input
directly as the first positional argument and the Pauli object will be
created just as with the label kwarg before.