Fix catch in grover tutorial for tera 0.24.0 compatibility
Qiskit Terra 0.24.0 had a bug that inadvertently changed the exception
type returned when using the PhaseOracle class without having tweedledum
installed locally. This commit changes the exception we catch in the
grovers tutorial using the phase oracle to be an ImportError instead of
the more specific MissingOptionalLibraryError exception. For terra
0.24.0 it is returning a ModuleNotFoundError instead of the correct
MissingOptionalLibraryError, this will be fixed in 0.24.1, but in the
meantime we can just use ImportError as it's the common parent class for
both exception types and will be compatible with both.