Switch to using Tox (#1459)
<!--
⚠️ If you do not respect this template, your pull request will be
closed.
⚠️ Your pull request title should be short detailed and understandable
for all.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary

Previously, there was no standardized way to build the docs locally.
While Tox isn't perfect, it aligns us with the rest of the Qiskit
ecosystem.

Switching to Tox is important here because it de-duplicates our
requirements, which were specified both in requirements-dev.txt and
again in azure-pipelines.yaml.

### Details and comments

This also fixes two Sphinx warnings due to bad config in `conf.py`.

I couldn't actually get NBSphinx to work on my mac due to import errors.
But this is still some forward progress, so I put up the PR. CI at least
will now use tox, and continue to produce a zip of the docs that CI
users can inspect.